Refactor: MeetingCommand 관련 메소드 도메인 서비스 계층으로 이동
