+++ src/main/java/com/boardgo/domain/meeting/service/facade/MeetingCommandFacade.java
@@ -7,12 +7,6 @@ import org.springframework.web.multipart.MultipartFile;
 public interface MeetingCommandFacade {
     Long create(MeetingCreateRequest meetingCreateRequest, MultipartFile imageFile, Long userId);
 
-    void incrementShareCount(Long meetingId);
-
-    void incrementViewCount(Long meetingId);
-
-    void updateCompleteMeetingState(Long meetingId);
-
     void deleteMeeting(Long meetingId, Long userId);
 
     void updateMeeting(
